LPS-193648 refactor AttachmentsUploadResponseHandler.java and CommerceMediaDefaultImageUploadFileEntryHandler.java
move and create inner class in UploadCommerceMediaDefaultImageMVCActionCommand.java and delete handler class
