+++ modules/apps/commerce/commerce-catalog-web/src/main/java/com/liferay/commerce/catalog/web/internal/portlet/action/UploadCommerceMediaDefaultImageMVCActionCommand.java
@@ -5,16 +5,44 @@
 
 package com.liferay.commerce.catalog.web.internal.portlet.action;
 
-import com.liferay.commerce.catalog.web.internal.upload.AttachmentsUploadResponseHandler;
-import com.liferay.commerce.catalog.web.internal.upload.CommerceMediaDefaultImageUploadFileEntryHandler;
+import com.liferay.commerce.product.configuration.AttachmentsConfiguration;
 import com.liferay.commerce.product.constants.CPPortletKeys;
+import com.liferay.commerce.product.exception.CPAttachmentFileEntryNameException;
+import com.liferay.commerce.product.exception.CPAttachmentFileEntrySizeException;
+import com.liferay.item.selector.ItemSelectorUploadResponseHandler;
+import com.liferay.petra.string.StringPool;
+import com.liferay.portal.configuration.metatype.bnd.util.ConfigurableUtil;
+import com.liferay.portal.kernel.exception.PortalException;
+import com.liferay.portal.kernel.json.JSONObject;
+import com.liferay.portal.kernel.json.JSONUtil;
+import com.liferay.portal.kernel.log.Log;
+import com.liferay.portal.kernel.log.LogFactoryUtil;
+import com.liferay.portal.kernel.model.Company;
 import com.liferay.portal.kernel.portlet.bridges.mvc.BaseMVCActionCommand;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCActionCommand;
+import com.liferay.portal.kernel.repository.model.FileEntry;
+import com.liferay.portal.kernel.servlet.ServletResponseConstants;
+import com.liferay.portal.kernel.theme.ThemeDisplay;
+import com.liferay.portal.kernel.upload.UploadPortletRequest;
+import com.liferay.portal.kernel.util.File;
+import com.liferay.portal.kernel.util.StringUtil;
+import com.liferay.portal.kernel.util.TempFileEntryUtil;
+import com.liferay.portal.kernel.util.WebKeys;
+import com.liferay.upload.UniqueFileNameProvider;
+import com.liferay.upload.UploadFileEntryHandler;
 import com.liferay.upload.UploadHandler;
+import com.liferay.upload.UploadResponseHandler;
+
+import java.io.IOException;
+import java.io.InputStream;
+
+import java.util.Map;
 
 import javax.portlet.ActionRequest;
 import javax.portlet.ActionResponse;
+import javax.portlet.PortletRequest;
 
+import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
 
@@ -23,6 +51,7 @@ import org.osgi.service.component.annotations.Reference;
  * @author Alessio Antonio Rendina
  */
 @Component(
+	configurationPid = "com.liferay.commerce.product.configuration.AttachmentsConfiguration",
 	property = {
 		"javax.portlet.name=" + CPPortletKeys.COMMERCE_CATALOGS,
 		"mvc.command.name=/commerce_catalogs/upload_commerce_media_default_image"
@@ -31,6 +60,162 @@ import org.osgi.service.component.annotations.Reference;
 )
 public class UploadCommerceMediaDefaultImageMVCActionCommand
 	extends BaseMVCActionCommand {
+
+	public class AttachmentsUploadResponseHandler
+		implements UploadResponseHandler {
+
+		@Override
+		public JSONObject onFailure(
+				PortletRequest portletRequest, PortalException portalException)
+			throws PortalException {
+
+			JSONObject jsonObject =
+				_itemSelectorUploadResponseHandler.onFailure(
+					portletRequest, portalException);
+
+			if (portalException instanceof CPAttachmentFileEntryNameException ||
+				portalException instanceof CPAttachmentFileEntrySizeException) {
+
+				String errorMessage = StringPool.BLANK;
+				int errorType = 0;
+
+				if (portalException instanceof
+						CPAttachmentFileEntryNameException) {
+
+					errorMessage = StringUtil.merge(
+						_attachmentsConfiguration.imageExtensions());
+
+					errorType =
+						ServletResponseConstants.SC_FILE_EXTENSION_EXCEPTION;
+				}
+				else if (portalException instanceof
+							CPAttachmentFileEntrySizeException) {
+
+					errorType = ServletResponseConstants.SC_FILE_SIZE_EXCEPTION;
+				}
+
+				jsonObject.put(
+					"error",
+					JSONUtil.put(
+						"errorType", errorType
+					).put(
+						"message", errorMessage
+					));
+			}
+			else {
+				throw portalException;
+			}
+
+			return jsonObject;
+		}
+
+		@Override
+		public JSONObject onSuccess(
+				UploadPortletRequest uploadPortletRequest, FileEntry fileEntry)
+			throws PortalException {
+
+			return _itemSelectorUploadResponseHandler.onSuccess(
+				uploadPortletRequest, fileEntry);
+		}
+
+	}
+
+	public class CommerceMediaDefaultImageUploadFileEntryHandler
+		implements UploadFileEntryHandler {
+
+		@Override
+		public FileEntry upload(UploadPortletRequest uploadPortletRequest)
+			throws IOException, PortalException {
+
+			String fileName = uploadPortletRequest.getFileName(_PARAMETER_NAME);
+
+			_validateFile(
+				fileName, uploadPortletRequest.getSize(_PARAMETER_NAME));
+
+			String contentType = uploadPortletRequest.getContentType(
+				_PARAMETER_NAME);
+
+			ThemeDisplay themeDisplay =
+				(ThemeDisplay)uploadPortletRequest.getAttribute(
+					WebKeys.THEME_DISPLAY);
+
+			try (InputStream inputStream = uploadPortletRequest.getFileAsStream(
+					_PARAMETER_NAME)) {
+
+				return _addFileEntry(
+					fileName, contentType, inputStream, themeDisplay);
+			}
+		}
+
+		private FileEntry _addFileEntry(
+				String fileName, String contentType, InputStream inputStream,
+				ThemeDisplay themeDisplay)
+			throws PortalException {
+
+			String uniqueFileName = _uniqueFileNameProvider.provide(
+				fileName, curFileName -> _exists(themeDisplay, curFileName));
+
+			Company company = themeDisplay.getCompany();
+
+			return TempFileEntryUtil.addTempFileEntry(
+				company.getGroupId(), themeDisplay.getUserId(),
+				_TEMP_FOLDER_NAME, uniqueFileName, inputStream, contentType);
+		}
+
+		private boolean _exists(ThemeDisplay themeDisplay, String curFileName) {
+			try {
+				FileEntry tempFileEntry = TempFileEntryUtil.getTempFileEntry(
+					themeDisplay.getScopeGroupId(), themeDisplay.getUserId(),
+					_TEMP_FOLDER_NAME, curFileName);
+
+				if (tempFileEntry != null) {
+					return true;
+				}
+
+				return false;
+			}
+			catch (PortalException portalException) {
+				if (_log.isDebugEnabled()) {
+					_log.debug(portalException);
+				}
+
+				return false;
+			}
+		}
+
+		private void _validateFile(String fileName, long size)
+			throws PortalException {
+
+			if ((_attachmentsConfiguration.imageMaxSize() > 0) &&
+				(size > _attachmentsConfiguration.imageMaxSize())) {
+
+				throw new CPAttachmentFileEntrySizeException();
+			}
+
+			String extension = _file.getExtension(fileName);
+
+			String[] imageExtensions =
+				_attachmentsConfiguration.imageExtensions();
+
+			for (String imageExtension : imageExtensions) {
+				if (StringPool.STAR.equals(imageExtension) ||
+					imageExtension.equals(StringPool.PERIOD + extension)) {
+
+					return;
+				}
+			}
+
+			throw new CPAttachmentFileEntryNameException(
+				"Invalid image for file name " + fileName);
+		}
+
+	}
+
+	@Activate
+	protected void activate(Map<String, Object> properties) {
+		_attachmentsConfiguration = ConfigurableUtil.createConfigurable(
+			AttachmentsConfiguration.class, properties);
+	}
 
 	@Override
 	protected void doProcessAction(
@@ -42,12 +227,31 @@ public class UploadCommerceMediaDefaultImageMVCActionCommand
 			_attachmentsUploadResponseHandler, actionRequest, actionResponse);
 	}
 
-	@Reference
-	private AttachmentsUploadResponseHandler _attachmentsUploadResponseHandler;
+	private static final String _PARAMETER_NAME = "imageSelectorFileName";
+
+	private static final String _TEMP_FOLDER_NAME =
+		CommerceMediaDefaultImageUploadFileEntryHandler.class.getName();
+
+	private static final Log _log = LogFactoryUtil.getLog(
+		UploadCommerceMediaDefaultImageMVCActionCommand.class);
+
+	private volatile AttachmentsConfiguration _attachmentsConfiguration;
+	private final AttachmentsUploadResponseHandler
+		_attachmentsUploadResponseHandler =
+			new AttachmentsUploadResponseHandler();
+	private final CommerceMediaDefaultImageUploadFileEntryHandler
+		_commerceMediaDefaultImageUploadFileEntryHandler =
+			new CommerceMediaDefaultImageUploadFileEntryHandler();
 
 	@Reference
-	private CommerceMediaDefaultImageUploadFileEntryHandler
-		_commerceMediaDefaultImageUploadFileEntryHandler;
+	private File _file;
+
+	@Reference
+	private ItemSelectorUploadResponseHandler
+		_itemSelectorUploadResponseHandler;
+
+	@Reference
+	private UniqueFileNameProvider _uniqueFileNameProvider;
 
 	@Reference
 	private UploadHandler _uploadHandler;
