--- modules/apps/commerce/commerce-catalog-web/src/main/java/com/liferay/commerce/catalog/web/internal/upload/AttachmentsUploadResponseHandler.java
@@ -1,108 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.commerce.catalog.web.internal.upload;
-
-import com.liferay.commerce.product.configuration.AttachmentsConfiguration;
-import com.liferay.commerce.product.exception.CPAttachmentFileEntryNameException;
-import com.liferay.commerce.product.exception.CPAttachmentFileEntrySizeException;
-import com.liferay.item.selector.ItemSelectorUploadResponseHandler;
-import com.liferay.petra.string.StringPool;
-import com.liferay.portal.configuration.metatype.bnd.util.ConfigurableUtil;
-import com.liferay.portal.kernel.exception.PortalException;
-import com.liferay.portal.kernel.json.JSONFactory;
-import com.liferay.portal.kernel.json.JSONObject;
-import com.liferay.portal.kernel.repository.model.FileEntry;
-import com.liferay.portal.kernel.servlet.ServletResponseConstants;
-import com.liferay.portal.kernel.upload.UploadPortletRequest;
-import com.liferay.portal.kernel.util.StringUtil;
-import com.liferay.upload.UploadResponseHandler;
-
-import java.util.Map;
-
-import javax.portlet.PortletRequest;
-
-import org.osgi.service.component.annotations.Activate;
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
-
-/**
- * @author Marco Leo
- */
-@Component(
-	configurationPid = "com.liferay.commerce.product.configuration.AttachmentsConfiguration",
-	service = AttachmentsUploadResponseHandler.class
-)
-public class AttachmentsUploadResponseHandler implements UploadResponseHandler {
-
-	@Override
-	public JSONObject onFailure(
-			PortletRequest portletRequest, PortalException portalException)
-		throws PortalException {
-
-		JSONObject jsonObject = _itemSelectorUploadResponseHandler.onFailure(
-			portletRequest, portalException);
-
-		if (portalException instanceof CPAttachmentFileEntryNameException ||
-			portalException instanceof CPAttachmentFileEntrySizeException) {
-
-			JSONObject errorJSONObject = _jsonFactory.createJSONObject();
-
-			String errorMessage = StringPool.BLANK;
-			int errorType = 0;
-
-			if (portalException instanceof CPAttachmentFileEntryNameException) {
-				errorMessage = StringUtil.merge(
-					_attachmentsConfiguration.imageExtensions());
-
-				errorType =
-					ServletResponseConstants.SC_FILE_EXTENSION_EXCEPTION;
-			}
-			else if (portalException instanceof
-						CPAttachmentFileEntrySizeException) {
-
-				errorType = ServletResponseConstants.SC_FILE_SIZE_EXCEPTION;
-			}
-
-			errorJSONObject.put(
-				"errorType", errorType
-			).put(
-				"message", errorMessage
-			);
-
-			jsonObject.put("error", errorJSONObject);
-		}
-		else {
-			throw portalException;
-		}
-
-		return jsonObject;
-	}
-
-	@Override
-	public JSONObject onSuccess(
-			UploadPortletRequest uploadPortletRequest, FileEntry fileEntry)
-		throws PortalException {
-
-		return _itemSelectorUploadResponseHandler.onSuccess(
-			uploadPortletRequest, fileEntry);
-	}
-
-	@Activate
-	protected void activate(Map<String, Object> properties) {
-		_attachmentsConfiguration = ConfigurableUtil.createConfigurable(
-			AttachmentsConfiguration.class, properties);
-	}
-
-	private volatile AttachmentsConfiguration _attachmentsConfiguration;
-
-	@Reference
-	private ItemSelectorUploadResponseHandler
-		_itemSelectorUploadResponseHandler;
-
-	@Reference
-	private JSONFactory _jsonFactory;
-
-}
